main

+++ assets/scripts/shopFunctions.js
@@ -10,6 +10,7 @@ const shopText = require('../db/shop/shop.json')
 
 async function getPack(bot, msg, packCount) {
   try {
+    console.log("функция начала свое выполнение")
     const userId = msg.from.id;
     const user = users.find((x) => x.username === msg.from.username);
 
@@ -22,8 +23,7 @@ async function getPack(bot, msg, packCount) {
     if (user.balance == null || isNaN(user.balance)) {
       user.balance = 0;
       fs.writeFileSync("assets/db/db.json", JSON.stringify(users, null, '\t'));
-    };
-    if (user.balance < totalCost) {
+    }else if (user.balance < totalCost) {
       user.balance = 0
       fs.writeFileSync("assets/db/db.json", JSON.stringify(users, null, '\t'))
       return bot.sendMessage(
@@ -36,25 +36,29 @@ async function getPack(bot, msg, packCount) {
     const shopMessage = (shopText.message|| "Текст не найден в магазине.").trim();
 
     for (let i = 0; i < packCount; i++) {
+      console.log("цикл начал свое выполнение")
       const randomCard = cards[Math.floor(Math.random() * cards.length)];
       const existingCard = user.inventory.find(
         (card) => card.cardName === randomCard.cardName
       );
-
-      if (existingCard in user.inventory && typeof existingCard.cardPower === "number") {
-        return (user.balance = randomCard.cardPower / 2);
+        console.log("фильрация карт завершена")
+      if (randomCard.cardName == existingCard.cardName) {
+        user.balance += randomCard.cardPower / 2;
+        fs.writeFileSync('assets/db/db.json', JSON.stringify(users, null, '\t'))
+        await bot.sendMessage(userId, `Вам выпала повторная карта ваш баланс составляет ${user.balance}`)
       } else if (typeof randomCard.cardPower === "number") {
         user.inventory.push(randomCard);
         openedCards.push(randomCard);
         user.balance += 0.5 * randomCard.cardPower;
+        fs.writeFileSync("assets/db/db.json", JSON.stringify(users, null, "\t"));
+
+        const photoMessage = `${shopMessage} Вы открыли ${packCount} паков и получили карты: ${openedCards.map((card) => card.name).join(", ")}. Новый баланс: ${user.balance}. сила карты ${openedCards.power}.`
+        await bot.sendPhoto(userId, randomCard.cardPhoto, {caption: photoMessage} );
       } else {
         return;
       }
     }
 
-    fs.writeFileSync("assets/db/db.json", JSON.stringify(users, null, "\t"));
-
-    await bot.sendPhoto(userId, `${shopMessage} Вы открыли ${packCount} паков и получили карты: ${openedCards.map((card) => card.name).join(", ")}. Новый баланс: ${user.balance}. сила карты ${openedCards.power}.`);
   } catch (error) {
     bot.sendMessage(msg.message.from.id, "Произошла ошибка при обработке вашего запроса.");
     throw error;
@@ -63,50 +67,36 @@ async function getPack(bot, msg, packCount) {
 
 async function getUniquePack(bot, msg) {
   try {
+    console.log("функция начала свое выполнение")
     const userId = msg.message.chat.id;
 
-    const user = users.find(user => user.username === msg.message.from.username);
+    const user = users.find(user => user.username === msg.from.username);
 
-    if (!user) {
-      return await bot.sendMessage(userId, "Пользователь не найден.");
-    }
-
-    const zeroDropChanceCards = cards.filter((card) => card.cardDropChance === 0);
+    const zeroDropChanceCards = cards.find((card) => card.cardDropChance === 0);
 
     const openedCards = [];
+    const shopMessage = (shopText.message|| "Текст не найден в магазине.").trim();
+
     let updatedBalance = user.balance || 0;
-
-    for (const randomCard of zeroDropChanceCards) {
-      const existingCardIndex = users.inventory.filter(
-        (card) => card.cardName === randomCard.name
-      );
-
-      if (existingCardIndex === -1) {
-        openedCards.push(randomCard);
-        updatedBalance -= randomCard.power;
-      } else {
-        const existingCard = users.inventory
-        const duplicateCard = { ...existingCard };
-        openedCards.push(duplicateCard);
-        updatedBalance -= duplicateCard.power;
-      }
-    }
-
-    if (isNaN(updatedBalance) || updatedBalance < 0) {
-      throw new Error("Некорректный баланс");
-    }
-
-    if (user.balance < updatedBalance) {
+    console.log("валидация юзера")
+    if (!user) {
+      return await bot.sendMessage(userId, "Пользователь не найден.");
+    }else if (user.balance < updatedBalance || isNaN(updatedBalance) || updatedBalance < 0) {
       return bot.sendMessage(userId, "У вас недостаточно баланса для открытия уникального пака.");
     }
-    user.inventory.push(openedCards);
+    console.log('валидация юзера прошла успешно')
 
-    fs.writeFileSync("assets/db/db.json", JSON.stringify(users, null, "\t"));
+    if (zeroDropChanceCards in user.inventory) {
+      user.inventory.push(zeroDropChanceCards);
+      fs.writeFileSync("assets/db/db.json", JSON.stringify(users, null, '\t'))
+      const photoMessage = `${shopMessage} Вы открыли уникальный пак и получили карту: ${user.inventory.filter((card) => card.cardName).join(", ")}. Новый баланс: ${user.balance}. сила карты ${zeroDropChanceCards.cardPower}.`
 
-    for (const card of openedCards) {
-      await bot.sendPhoto(userId, card.cardPhoto, {
-        caption: `🦠 ${card.cardName}\n\n💬 ${user.username}, поздравляем, вы получили карту героя ${card.cardName}!\n🎭 Класс: ${card.cardSection}\n🔮 Редкость: ${card.cardRarity}\nАтака: ${card.cardPower}\n❤️ Защита: ${card.cardDeffence}\n➖➖➖➖➖➖➖\n🃏 Кол-во оставшихся токенов: ${user.balance}`,
-      });
+      await bot.sendMessage(userId, "Вам выпала повторная карта ваш баланс составляет: " + user.balance)
+
+    }else{
+      user.inventory.push(zeroDropChanceCards);
+      fs.writeFileSync("assets/db/db.json", JSON.stringify(users, null, "\t"));
+      await bot.sendPhoto(userId, zeroDropChanceCards.cardPhoto, {caption: photoMessage} );
     }
   } catch (error) {
     bot.sendMessage(msg.message.chat.id,"Произошла ошибка при обработке вашего запроса.");
